Add a test case that my design can't handle.
I was skeptical that my design was doomed to fail for all inputs. However,
I think I've found the nail-in-the-coffin test case for why it will always
fail: Any matching against a regex will require us to consume input to decide
what branch to take. However, we cannot be certain that taking that branch will
pan out until we parse the entirety of the rest of the string. There's no way
around that, meaning that we have worst-case exponential complexity. Now I'll
start researching a way around this problem.